requirements.txt
pyusb~=1.3.1
